refactor AbstractDialog gui scaling
