+++ megamek/src/megamek/client/ui/dialogs/BVDisplayDialog.java
@@ -48,7 +48,6 @@ public class BVDisplayDialog extends AbstractDialog {
     protected void finalizeInitialization() throws Exception {
         super.finalizeInitialization();
         setTitle(getTitle() + " (" + entity.getShortName() + ")");
-        adaptToGUIScale();
         pack();
         Dimension screenSize = UIUtil.getScaledScreenSize(this);
         setSize(new Dimension(getSize().width, Math.min(getHeight(), (int) (screenSize.getHeight() * 0.8))));
@@ -87,9 +86,5 @@ public class BVDisplayDialog extends AbstractDialog {
     private void copyToClipboard(String reportString) {
         Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
         clipboard.setContents(new StringSelection(reportString), null);
-    }
-
-    private void adaptToGUIScale() {
-        UIUtil.adjustDialog(this,  UIUtil.FONT_SCALE1);
     }
 }
