+++ megamek/src/megamek/client/ui/dialogs/BotConfigDialog.java
@@ -129,7 +129,6 @@ public class BotConfigDialog extends AbstractButtonDialog implements ActionListe
         updatePresets();
         initialize();
         updateDialogFields();
-        adaptToGUIScale();
     }
     
     @Override
@@ -834,9 +833,5 @@ public class BotConfigDialog extends AbstractButtonDialog implements ActionListe
     @Override
     public void stateChanged(ChangeEvent e) {
         updateEnabledStates();
-    }
-
-    private void adaptToGUIScale() {
-        UIUtil.adjustDialog(this,  UIUtil.FONT_SCALE1);
     }
 }
