+++ megamek/src/megamek/client/ui/swing/GameOptionsDialog.java
@@ -33,8 +33,6 @@ import javax.xml.parsers.DocumentBuilder;
 
 import megamek.client.ui.baseComponents.AbstractButtonDialog;
 import megamek.client.ui.swing.util.UIUtil;
-import megamek.common.preference.IPreferenceChangeListener;
-import megamek.common.preference.PreferenceChangeEvent;
 import org.w3c.dom.Document;
 import org.w3c.dom.NodeList;
 
@@ -48,7 +46,7 @@ import static megamek.client.ui.swing.util.UIUtil.*;
 import static megamek.client.ui.Messages.*;
 
 /** Responsible for displaying the current game options and allowing the user to change them. */
-public class GameOptionsDialog extends AbstractButtonDialog implements ActionListener, DialogOptionListener, IPreferenceChangeListener {
+public class GameOptionsDialog extends AbstractButtonDialog implements ActionListener, DialogOptionListener {
 
     private ClientGUI clientGui;
     private JFrame frame;
@@ -141,9 +139,6 @@ public class GameOptionsDialog extends AbstractButtonDialog implements ActionLis
         mainPanel.add(Box.createVerticalStrut(5));
         mainPanel.add(panButtons);
 
-        adaptToGUIScale();
-        GUIPreferences.getInstance().addPreferenceChangeListener(this);
-
         return mainPanel;
     }
 
@@ -165,8 +160,6 @@ public class GameOptionsDialog extends AbstractButtonDialog implements ActionLis
         panButtons.add(butDefaults);
         panButtons.add(butSave);
         panButtons.add(butLoad);
-
-        adaptToGUIScale();
 
         return panButtons;
     }
@@ -912,17 +905,5 @@ public class GameOptionsDialog extends AbstractButtonDialog implements ActionLis
      */
     public boolean isEditable() {
         return editable;
-    }
-
-    private void adaptToGUIScale() {
-        UIUtil.adjustDialog(this, UIUtil.FONT_SCALE1);
-    }
-
-    @Override
-    public void preferenceChange(PreferenceChangeEvent e) {
-        // Update the text size when the GUI scaling changes
-        if (e.getName().equals(GUIPreferences.GUI_SCALE)) {
-            adaptToGUIScale();
-        }
     }
 }
