+++ megamek/src/megamek/client/ui/swing/dialog/AdvancedSearchDialog2.java
@@ -21,7 +21,6 @@ package megamek.client.ui.swing.dialog;
 import megamek.client.ui.baseComponents.AbstractButtonDialog;
 import megamek.client.ui.swing.unitSelector.ASAdvancedSearchPanel;
 import megamek.client.ui.swing.unitSelector.TWAdvancedSearchPanel;
-import megamek.client.ui.swing.util.UIUtil;
 
 import javax.swing.*;
 import javax.swing.border.EmptyBorder;
@@ -51,9 +50,6 @@ public class AdvancedSearchDialog2 extends AbstractButtonDialog {
     @Override
     public void setVisible(boolean b) {
         alphaStrikeTab.saveValues();
-        if (b) {
-            adaptToGUIScale();
-        }
         super.setVisible(b);
     }
 
@@ -97,8 +93,9 @@ public class AdvancedSearchDialog2 extends AbstractButtonDialog {
         return totalWarTab;
     }
 
-    private void adaptToGUIScale() {
-        UIUtil.adjustDialog(this, UIUtil.FONT_SCALE1);
+    @Override
+    public void adaptToGUIScale() {
+        super.adaptToGUIScale();
         totalWarTab.adaptToGUIScale();
     }
 }
